# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NapsterTec Intelligence Engine",
    "description": "Enterprise multi-modal intelligence and monitoring platform.",
    "version": "1.4.0"
  },
  "paths": {
...
    "/api/v1/memory/conversations": {
      "post": {
        "tags": [
          "Memory"
        ],
        "summary": "Create Conversation",
        "operationId": "create_conversation_api_v1_memory_conversations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Memory"
        ],
        "summary": "List Conversations",
        "operationId": "list_conversations_api_v1_memory_conversations_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "local_user",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationResponse"
                  },
                  "title": "Response List Conversations Api V1 Memory Conversations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/conversations/{conversation_id}/messages": {
      "post": {
        "tags": [
          "Memory"
        ],
        "summary": "Add Message",
        "operationId": "add_message_api_v1_memory_conversations__conversation_id__messages_post",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Memory"
        ],
        "summary": "Get Messages",
        "operationId": "get_messages_api_v1_memory_conversations__conversation_id__messages_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Get Messages Api V1 Memory Conversations  Conversation Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/analytics/dashboard": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Dashboard Aggregated",
        "operationId": "get_dashboard_aggregated_analytics_dashboard_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/providers": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Providers",
        "operationId": "get_providers_analytics_providers_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderStat"
                  },
                  "title": "Response Get Providers Analytics Providers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/skills": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Skills",
        "operationId": "get_skills_analytics_skills_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillStat"
                  },
                  "title": "Response Get Skills Analytics Skills Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/usage": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Usage",
        "operationId": "get_usage_analytics_usage_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyUsageStat"
                  },
                  "title": "Response Get Usage Analytics Usage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/latency": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Latency",
        "operationId": "get_latency_analytics_latency_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/fallbacks": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Fallbacks",
        "operationId": "get_fallbacks_analytics_fallbacks_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FallbackStat"
                  },
                  "title": "Response Get Fallbacks Analytics Fallbacks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationCreate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "default": "New Chat"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "default": "local_user"
          }
        },
        "type": "object",
        "title": "ConversationCreate"
      },
      "ConversationResponse": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "default": "New Chat"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "default": "local_user"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "ConversationResponse"
      },
      "DailyUsageStat": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "requests": {
            "type": "integer",
            "title": "Requests"
          },
          "tokens": {
            "type": "integer",
            "title": "Tokens"
          }
        },
        "type": "object",
        "required": [
          "date",
          "requests",
          "tokens"
        ],
        "title": "DailyUsageStat"
      },
      "DashboardResponse": {
        "properties": {
          "total_requests": {
            "type": "integer",
            "title": "Total Requests"
          },
          "total_fallbacks": {
            "type": "integer",
            "title": "Total Fallbacks"
          },
          "average_response_time_ms": {
            "type": "number",
            "title": "Average Response Time Ms"
          },
          "success_rate": {
            "type": "number",
            "title": "Success Rate"
          },
          "active_providers": {
            "type": "integer",
            "title": "Active Providers"
          },
          "provider_breakdown": {
            "items": {
              "$ref": "#/components/schemas/ProviderStat"
            },
            "type": "array",
            "title": "Provider Breakdown"
          },
          "skill_usage": {
            "items": {
              "$ref": "#/components/schemas/SkillStat"
            },
            "type": "array",
            "title": "Skill Usage"
          },
          "daily_trend": {
            "items": {
              "$ref": "#/components/schemas/DailyUsageStat"
            },
            "type": "array",
            "title": "Daily Trend"
          }
        },
        "type": "object",
        "required": [
          "total_requests",
          "total_fallbacks",
          "average_response_time_ms",
          "success_rate",
          "active_providers",
          "provider_breakdown",
          "skill_usage",
          "daily_trend"
        ],
        "title": "DashboardResponse"
      },
...
      "FallbackStat": {
        "properties": {
          "original_provider": {
            "type": "string",
            "title": "Original Provider"
          },
          "fallback_provider": {
            "type": "string",
            "title": "Fallback Provider"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "original_provider",
          "fallback_provider",
          "count"
        ],
        "title": "FallbackStat"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageCreate": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tokens_used": {
            "type": "integer",
            "title": "Tokens Used",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "MessageCreate"
      },
      "MessageResponse": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tokens_used": {
            "type": "integer",
            "title": "Tokens Used",
            "default": 0
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content",
          "id",
          "conversation_id",
          "created_at"
        ],
        "title": "MessageResponse"
      },
      "ProviderStat": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "request_count": {
            "type": "integer",
            "title": "Request Count"
          },
          "average_latency_ms": {
            "type": "number",
            "title": "Average Latency Ms"
          },
          "error_rate": {
            "type": "number",
            "title": "Error Rate"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "request_count",
          "average_latency_ms",
          "error_rate"
        ],
        "title": "ProviderStat"
      },
...
      "SkillStat": {
        "properties": {
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "usage_count": {
            "type": "integer",
            "title": "Usage Count"
          }
        },
        "type": "object",
        "required": [
          "skill",
          "usage_count"
        ],
        "title": "SkillStat"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}